moved navbar styling file and  fixed navbar code
