+++ .env
@@ -1,0 +1,7 @@
+NEXT_PUBLIC_FIREBASE_API_KEY=AIzaSyCCaOjWj_UKwIUkdlu0RGlJLXcpPn7a3Yo
+NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN=siren-6099f.firebaseapp.com
+NEXT_PUBLIC_FIREBASE_PROJECT_ID=siren-6099f
+NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET=siren-6099f.appspot.com
+NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID=842040742509
+NEXT_PUBLIC_FIREBASE_APP_ID=1:842040742509:web:5925513120806909e46564
+NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID=G-F1BP31WZ7R
