+++ components/Navbar.module.css
@@ -1,0 +1,8 @@
+.bar {
+  display: flex; 
+  width: 100%;
+  height: 5%; 
+  /* color will change the color of the text */
+  /* try using 'background-color' instead! */
+  color: #003153; 
+}
