+++ components/Navbar.tsx
@@ -1,20 +1,16 @@
 import Link from 'next/link'
-import styles from '../styles/Navbar.module.css'
+import styles from './Navbar.module.css'
 
 export const Navbar = () => {   
     return (
-        <nav>
-            <div className={styles.bar}>
-                <Link href="/intake-dashboard">
-                    <a>Intake Dashboard</a>
-                </Link>
-            </div>
+        <nav className={styles.bar}>
+            <Link href="/intake-dashboard">
+                <a>Intake Dashboard</a>
+            </Link>
         
-            <div className={styles.bar}>
-                <Link href="/calendar">
-                    <a>Calendar</a>
-                </Link>
-            </div>
+            <Link href="/calendar">
+                <a>Calendar</a>
+            </Link>
         </nav>
     )
 }
