+++ components/layout.tsx
@@ -1,4 +1,4 @@
-import { Navbar } from '../components/Navbar'
+import { Navbar } from './Navbar'
 
 export default function Layout({ children }) {
   return (
