+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import Head from 'next/head'
 import styles from '../styles/Home.module.css'
 import firebase from '../firebase/clientApp'
-import { useCollection } from "react-firebase-hooks/firestore";
+import { useCollection } from "react-firebase-hooks/firestore"
 
 export default function Home() {
