--- styles/Navbar.module.css
@@ -1,7 +1,0 @@
-.bar {
-    padding: 5px;
-    height: 100vh;
-    width: 10%;
-    color: #003153;
-    margin: 5px ;
-  }
